+++ test/specs/events/hidden.spec.js
@@ -51,4 +51,41 @@ describe('hidden (event)', () => {
 
     viewer.show();
   });
+
+  it('should have expected properties in `event.detail`', (done) => {
+    const image = window.createImage();
+
+    image.addEventListener('hidden', (event) => {
+      const { detail } = event;
+
+      expect(detail).to.be.an('object').that.has.all.keys('originalEvent');
+      expect(detail.originalEvent).to.be.null;
+      done();
+    });
+
+    const viewer = new Viewer(image, {
+      shown() {
+        viewer.hide(true);
+      },
+    });
+
+    viewer.show();
+  });
+
+  it('should pass the original event to `event.detail.originalEvent` when triggered by a click', (done) => {
+    const image = window.createImage();
+
+    image.addEventListener('hidden', (event) => {
+      expect(event.detail.originalEvent).to.be.an.instanceOf(window.MouseEvent);
+      done();
+    });
+
+    const viewer = new Viewer(image, {
+      shown() {
+        viewer.button.click();
+      },
+    });
+
+    viewer.show();
+  });
 });