+++ test/specs/events/play.spec.js
@@ -37,4 +37,45 @@ describe('play (event)', () => {
     viewer = new Viewer(image);
     viewer.show();
   });
+
+  it('should have expected properties in `event.detail`', (done) => {
+    const image = window.createImage();
+    let viewer;
+
+    image.addEventListener('viewed', () => {
+      viewer.play();
+    });
+
+    image.addEventListener('play', (event) => {
+      const { detail } = event;
+
+      expect(detail).to.be.an('object').that.has.all.keys('originalEvent');
+      expect(detail.originalEvent).to.be.null;
+      event.preventDefault();
+      viewer.hide(true);
+      done();
+    });
+
+    viewer = new Viewer(image);
+    viewer.show();
+  });
+
+  it('should pass the original event to `event.detail.originalEvent` when triggered by a click', (done) => {
+    const image = window.createImage();
+    let viewer;
+
+    image.addEventListener('viewed', () => {
+      viewer.toolbar.querySelector('.viewer-play').click();
+    });
+
+    image.addEventListener('play', (event) => {
+      expect(event.detail.originalEvent).to.be.an.instanceOf(window.MouseEvent);
+      event.preventDefault();
+      viewer.hide(true);
+      done();
+    });
+
+    viewer = new Viewer(image);
+    viewer.show();
+  });
 });