+++ test/specs/events/rotate.spec.js
@@ -29,9 +29,10 @@ describe('rotate (event)', () => {
     image.addEventListener('rotate', (event) => {
       const { detail } = event;
 
-      expect(detail).to.be.an('object').that.has.all.keys('degree', 'oldDegree');
+      expect(detail).to.be.an('object').that.has.all.keys('degree', 'oldDegree', 'originalEvent');
       expect(detail.degree).to.be.a('number');
       expect(detail.oldDegree).to.be.a('number');
+      expect(detail.originalEvent).to.be.null;
       event.preventDefault();
       viewer.hide(true);
       done();
@@ -79,5 +80,24 @@ describe('rotate (event)', () => {
     viewer = new Viewer(image, {
       inline: true,
     });
+  });
+
+  it('should pass the original event to `event.detail.originalEvent` when triggered by a click', (done) => {
+    const image = window.createImage();
+    let viewer;
+
+    image.addEventListener('viewed', () => {
+      viewer.toolbar.querySelector('.viewer-rotate-left').click();
+    });
+
+    image.addEventListener('rotate', (event) => {
+      expect(event.detail.originalEvent).to.be.an.instanceOf(window.MouseEvent);
+      event.preventDefault();
+      viewer.hide(true);
+      done();
+    });
+
+    viewer = new Viewer(image);
+    viewer.show();
   });
 });