+++ test/specs/events/scale.spec.js
@@ -29,11 +29,12 @@ describe('scale (event)', () => {
     image.addEventListener('scale', (event) => {
       const { detail } = event;
 
-      expect(detail).to.be.an('object').that.has.all.keys('scaleX', 'scaleY', 'oldScaleX', 'oldScaleY');
+      expect(detail).to.be.an('object').that.has.all.keys('scaleX', 'scaleY', 'oldScaleX', 'oldScaleY', 'originalEvent');
       expect(detail.scaleX).to.be.a('number');
       expect(detail.scaleY).to.be.a('number');
       expect(detail.oldScaleX).to.be.a('number');
       expect(detail.oldScaleY).to.be.a('number');
+      expect(detail.originalEvent).to.be.null;
       event.preventDefault();
       viewer.hide(true);
       done();
@@ -81,5 +82,24 @@ describe('scale (event)', () => {
     viewer = new Viewer(image, {
       inline: true,
     });
+  });
+
+  it('should pass the original event to `event.detail.originalEvent` when triggered by a click', (done) => {
+    const image = window.createImage();
+    let viewer;
+
+    image.addEventListener('viewed', () => {
+      viewer.toolbar.querySelector('.viewer-flip-horizontal').click();
+    });
+
+    image.addEventListener('scale', (event) => {
+      expect(event.detail.originalEvent).to.be.an.instanceOf(window.MouseEvent);
+      event.preventDefault();
+      viewer.hide(true);
+      done();
+    });
+
+    viewer = new Viewer(image);
+    viewer.show();
   });
 });