+++ test/specs/events/show.spec.js
@@ -41,4 +41,36 @@ describe('show (event)', () => {
       inline: true,
     });
   });
+
+  it('should have expected properties in `event.detail`', (done) => {
+    const image = window.createImage();
+
+    image.addEventListener('show', (event) => {
+      const { detail } = event;
+
+      expect(detail).to.be.an('object').that.has.all.keys('originalEvent');
+      expect(detail.originalEvent).to.be.null;
+      event.preventDefault();
+      done();
+    });
+
+    const viewer = new Viewer(image);
+
+    viewer.show();
+  });
+
+  it('should pass the original event to `event.detail.originalEvent` when triggered by a click', (done) => {
+    const image = window.createImage();
+    let viewer;
+
+    image.addEventListener('show', (event) => {
+      expect(event.detail.originalEvent).to.be.an.instanceOf(window.MouseEvent);
+      event.preventDefault();
+      viewer.hide(true);
+      done();
+    });
+
+    viewer = new Viewer(image);
+    image.click();
+  });
 });