+++ test/specs/events/viewed.spec.js
@@ -20,10 +20,11 @@ describe('viewed (event)', () => {
     image.addEventListener('viewed', (event) => {
       const { detail } = event;
 
-      expect(detail).to.be.an('object').that.has.all.keys('image', 'index', 'originalImage');
+      expect(detail).to.be.an('object').that.has.all.keys('image', 'index', 'originalImage', 'originalEvent');
       expect(detail.image.src).to.equal(image.src);
       expect(detail.index).to.equal(0);
       expect(detail.originalImage).to.equal(image);
+      expect(detail.originalEvent).to.be.null;
       viewer.hide(true);
       done();
     });
@@ -59,5 +60,20 @@ describe('viewed (event)', () => {
 
     viewer = new Viewer(image);
     viewer.show();
+  });
+
+  it('should pass the original event to `event.detail.originalEvent` when triggered by a click', (done) => {
+    const image = window.createImage();
+    let viewer;
+
+    image.addEventListener('viewed', (event) => {
+      expect(event.detail.originalEvent).to.be.an.instanceOf(window.MouseEvent);
+      event.preventDefault();
+      viewer.hide(true);
+      done();
+    });
+
+    viewer = new Viewer(image);
+    image.click();
   });
 });