+++ test/specs/options/rotate.spec.js
@@ -34,9 +34,10 @@ describe('rotate (option)', () => {
       rotate(event) {
         const { detail } = event;
 
-        expect(detail).to.be.an('object').that.has.all.keys('degree', 'oldDegree');
+        expect(detail).to.be.an('object').that.has.all.keys('degree', 'oldDegree', 'originalEvent');
         expect(detail.degree).to.be.a('number');
         expect(detail.oldDegree).to.be.a('number');
+        expect(detail.originalEvent).to.be.null;
         event.preventDefault();
         viewer.hide(true);
         done();