+++ test/specs/options/scaled.spec.js
@@ -33,11 +33,12 @@ describe('scaled (option)', () => {
       scaled(event) {
         const { detail } = event;
 
-        expect(detail).to.be.an('object').that.has.all.keys('scaleX', 'scaleY', 'oldScaleX', 'oldScaleY');
+        expect(detail).to.be.an('object').that.has.all.keys('scaleX', 'scaleY', 'oldScaleX', 'oldScaleY', 'originalEvent');
         expect(detail.scaleX).to.be.a('number');
         expect(detail.scaleY).to.be.a('number');
         expect(detail.oldScaleX).to.be.a('number');
         expect(detail.oldScaleY).to.be.a('number');
+        expect(detail.originalEvent).to.be.null;
         viewer.hide(true);
         done();
       },