+++ test/specs/options/view.spec.js
@@ -26,10 +26,11 @@ describe('view (option)', () => {
     const image = window.createImage();
     const viewer = new Viewer(image, {
       view(event) {
-        expect(event.detail).to.be.an('object').that.has.all.keys('image', 'index', 'originalImage');
+        expect(event.detail).to.be.an('object').that.has.all.keys('image', 'index', 'originalImage', 'originalEvent');
         expect(event.detail.image.src).to.equal(image.src);
         expect(event.detail.index).to.equal(0);
         expect(event.detail.originalImage).to.equal(image);
+        expect(event.detail.originalEvent).to.be.null;
       },
 
       viewed() {